requirements.txt
# FastAPI and web server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Core ML dependencies (compatible versions)
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.30.2
numpy==1.24.3

# Optional ML dependencies
faiss-cpu==1.7.4
openai==1.6.1
groq==0.4.1
langchain==0.1.0
langchain-core==0.1.12
langchain-groq==0.0.1

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
chardet==5.2.0

# Database and storage
aiosqlite==0.19.0

# Utilities and async support
aiofiles==23.2.1
httpx==0.25.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Environment and configuration
python-dotenv==1.0.0

# Logging
structlog==23.2.0
